# src/main/java/src/main/java/com/viafluvial/srvusuario/adapters/in/web/api/AdminsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.viafluvial.srvusuario.adapters.in.web.api;

import com.viafluvial.srvusuario.adapters.in.web.dto.AdminDTOApi;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-02-07T21:43:37.860179204Z[Etc/UTC]", comments = "Generator version: 7.6.0")
@Validated
@Tag(name = "Admins", description = "the Admins API")
public interface AdminsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /admins : Create admin
     *
     * @param adminDTOApi  (required)
     * @return Admin created (status code 201)
     *         or Invalid request (status code 400)
     */
    @Operation(
        operationId = "createAdmin",
        summary = "Create admin",
        tags = { "Admins" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Admin created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AdminDTOApi.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/admins",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<AdminDTOApi> createAdmin(
        @Parameter(name = "AdminDTOApi", description = "", required = true) @Valid @RequestBody AdminDTOApi adminDTOApi
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"role\" : \"SUPER_ADMIN\", \"permissions\" : \"permissions\", \"employeeId\" : \"employeeId\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"department\" : \"department\", \"userId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /admins/{id} : Get admin by id
     *
     * @param id  (required)
     * @return Admin found (status code 200)
     *         or Admin not found (status code 404)
     */
    @Operation(
        operationId = "getAdminById",
        summary = "Get admin by id",
        tags = { "Admins" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Admin found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AdminDTOApi.class))
            }),
            @ApiResponse(responseCode = "404", description = "Admin not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/admins/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<AdminDTOApi> getAdminById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"role\" : \"SUPER_ADMIN\", \"permissions\" : \"permissions\", \"employeeId\" : \"employeeId\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"department\" : \"department\", \"userId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /admins/user/{userId} : Get admin by user id
     *
     * @param userId  (required)
     * @return Admin found (status code 200)
     *         or Admin not found (status code 404)
     */
    @Operation(
        operationId = "getAdminByUserId",
        summary = "Get admin by user id",
        tags = { "Admins" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Admin found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AdminDTOApi.class))
            }),
            @ApiResponse(responseCode = "404", description = "Admin not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/admins/user/{userId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<AdminDTOApi> getAdminByUserId(
        @Parameter(name = "userId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("userId") UUID userId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"role\" : \"SUPER_ADMIN\", \"permissions\" : \"permissions\", \"employeeId\" : \"employeeId\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"department\" : \"department\", \"userId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /admins/{id} : Update admin
     *
     * @param id  (required)
     * @param adminDTOApi  (required)
     * @return Admin updated (status code 200)
     *         or Admin not found (status code 404)
     */
    @Operation(
        operationId = "updateAdmin",
        summary = "Update admin",
        tags = { "Admins" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Admin updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AdminDTOApi.class))
            }),
            @ApiResponse(responseCode = "404", description = "Admin not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/admins/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<AdminDTOApi> updateAdmin(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id,
        @Parameter(name = "AdminDTOApi", description = "", required = true) @Valid @RequestBody AdminDTOApi adminDTOApi
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"role\" : \"SUPER_ADMIN\", \"permissions\" : \"permissions\", \"employeeId\" : \"employeeId\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"department\" : \"department\", \"userId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}