# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ViaFluvial User Service API
  version: 1.0.0
servers:
  - url: /api/v1
...
paths:
...
  /admins:
    post:
      tags: [Admins]
      summary: Create admin
      operationId: createAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminDTO'
      responses:
        '201':
          description: Admin created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDTO'
        '400':
          description: Invalid request
  /admins/{id}:
    get:
      tags: [Admins]
      summary: Get admin by id
      operationId: getAdminById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Admin found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDTO'
        '404':
          description: Admin not found
    put:
      tags: [Admins]
      summary: Update admin
      operationId: updateAdmin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminDTO'
      responses:
        '200':
          description: Admin updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDTO'
        '404':
          description: Admin not found
  /admins/user/{userId}:
    get:
      tags: [Admins]
      summary: Get admin by user id
      operationId: getAdminByUserId
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Admin found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDTO'
        '404':
          description: Admin not found
...
components:
  schemas:
...
    AdminRole:
      type: string
      enum: [SUPER_ADMIN, ADMIN, FINANCIAL, SUPPORT]
...
    AdminDTO:
      type: object
      required: [userId, role]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/AdminRole'
        permissions:
          type: string
        department:
          type: string
        employeeId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...